Extract tool_display.rs from turn.rs (parent #680)
Move the per-iteration progress line rendering cluster out of turn.rs
into a new src/agent/loop_run/tool_display.rs sibling module:

- ProgressDisplay struct
- progress_path_display (pub(super))
- tool_display (pub(super), entry point)
- tool_display_str_arg / tool_display_write / tool_display_edit /
  tool_display_preview_note / tool_display_read /
  tool_display_plan_read / tool_display_workspace_read /
  tool_display_bash / tool_display_search / tool_display_default
- summarize_plan_read / read_line_suffix / text_preview /
  read_preview / compact_progress_path (private helpers)

All items are pub(super), no facade re-export (DR3-001). Callers updated:
- turn.rs: imports `progress_path_display` (cfg(test): `tool_display`)
- actor_loop_flow.rs: `tool_display` now comes from `super::tool_display`
- scaffold_pipeline.rs: `progress_path_display` now from `super::tool_display`

Cleaned unused turn.rs imports (PlanStage, sanitize_for_progress, truncate
are now test-only after the migration).

turn.rs: 27,739 -> 27,394 (-345 LOC)
Cumulative from 39,489: -12,095 LOC (-30.6%)

Verification:
- cargo fmt --check
- cargo clippy --lib --all-targets -- -D warnings
- cargo test --lib (3,114 passed)
