+++ src/agent/loop_run/actor_loop_flow.rs
@@ -66,13 +66,13 @@ use super::scaffold_pipeline::PlanExplorationKey;
 use super::spinner::Spinner;
 use super::success::DETERMINISTIC_CONTENT_FALLBACK_TAG;
 use super::summary::{ExitReason, LoopResult, LoopStats};
+use super::tool_display::tool_display;
 use super::tool_history::focused_edit_target_already_read;
 use super::tool_history::is_plan_file_tool_call;
 use super::tool_policy::EffectiveToolPolicy;
 use super::turn::{
     LOG_ARGS_MAX_CHARS, PLAN_REPEATED_EXPLORATION_BLOCK_THRESHOLD, join_sections_for_progress,
-    plan_section_body_for_progress, plan_sections_with_content, tool_display,
-    write_stdout_rendered,
+    plan_section_body_for_progress, plan_sections_with_content, write_stdout_rendered,
 };
 use crate::agent::prompting;
 use crate::session::compact::approximate_token_count;
