+++ src/agent/loop_run/scaffold_pipeline.rs
@@ -52,15 +52,16 @@ use super::quality::{
     package_json_with_requested_port, react_dev_wrapper_for_requested_port,
 };
 use super::task_contract::{ArtifactRole, CompletionDecision};
+use super::tool_display::progress_path_display;
 use super::tool_history::{
     focused_edit_target_already_read, has_successful_non_plan_repo_edit, latest_user_turn_slice,
 };
 use super::turn::{
     WrittenScaffoldArtifacts, current_file_hash_for_relative_path, extract_filename_with_suffix,
     last_read_tool_path, latest_turn_preferred_read_edit_target, meaningful_workspace_files,
-    progress_path_display, sha256_hex, should_fallback_plan_model_after_timeout,
-    should_materialize_plan_after_timeout, should_materialize_plan_after_tool_call_format_error,
-    tool_result_failed, workspace_appears_empty, write_stdout_rendered,
+    sha256_hex, should_fallback_plan_model_after_timeout, should_materialize_plan_after_timeout,
+    should_materialize_plan_after_tool_call_format_error, tool_result_failed,
+    workspace_appears_empty, write_stdout_rendered,
 };
 use crate::agent::prompting;
 use crate::agent::recovery;
